에너미가 움직이는 방향이 바뀌면 이미지 방향도 바뀐다

+++ enemy.h
@@ -16,7 +16,7 @@ private:
 
 	//bool _isLeft;
 	//bool _isAttack;
-	int _speed;
+	float _speed;
 
 	ENEMY_STATE _state;
